requirements.txt
-f https://storage.googleapis.com/jax-releases/libtpu_releases.html
--extra-index-url https://download.pytorch.org/whl/cpu

jax[tpu]==0.4.37
jaxlib==0.4.36
flax==0.10.4
optax==0.2.5
orbax-checkpoint==0.11.0
numpy<2
ml-dtypes==0.5.4
tensorstore==0.1.78

torch==2.4.0+cpu
torchvision==0.19.0+cpu

diffusers==0.37.0
transformers==5.3.0
huggingface_hub==1.7.1

absl-py==2.4.0
chex==0.1.90
einops==0.8.2
etils==1.13.0
fsspec==2026.2.0
GitPython==3.1.46
hf-xet==1.4.2
humanize==4.15.0
ml-collections==1.1.0
msgpack==1.1.2
networkx==3.4.2
Pillow==12.1.1
protobuf==6.33.5
PyYAML==6.0.3
regex==2026.2.28
safetensors==0.7.0
scipy==1.15.3
simplejson==3.20.2
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
wandb==0.25.1
